backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "arc-backend"
version = "0.1.0"
description = "Arc — Archenemies Deep Agent backend"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Deep Agents SDK (batteries-included agent harness)
    "deepagents>=0.4.12",
    # LangGraph — stateful agent orchestration
    "langgraph>=1.1.3",
    # LangChain — model integrations and tool primitives
    "langchain>=0.3",
    "langchain-core>=0.3",
    # LLM provider integrations (install what you need)
    "langchain-openai>=0.3",
    "langchain-anthropic>=0.3",
    "langchain-google-genai>=2.0",
    "langchain-openrouter>=0.2",
    # Web search tool for the research sub-agent
    "langchain-tavily>=0.2",
    # MCP tool adapters
    "langchain-mcp-adapters>=0.1",
    # API server
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    # Async HTTP client (for SSE streaming)
    "httpx>=0.28",
    # Environment variable loading
    "python-dotenv>=1.0",
    # PostgreSQL checkpointer and store (production persistence)
    "langgraph-checkpoint-postgres>=2.0",
    # System monitoring
    "psutil>=6.0",
    # Type stubs
    "typing-extensions>=4.12",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.9",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
